[Decomp][Unit] Cohesion split: extract melee combat resolution to UnitCombat.cpp
